refactor: (bridge) extract out pure functions
