+++ packages/discord-bridge/src/discord/backfill.ts
@@ -8,7 +8,9 @@ import {
 import { GuildContext } from "../types.js";
 import { tracer, setDiscordAttrs, recordError } from "../tracing.js";
 import { DISCORD_EXTENSION_KEYS } from "../roomy/subscription.js";
-import { getRoomKey, fingerprint, syncDiscordReactionToRoomy } from "../roomy/to.js";
+import { syncDiscordReactionToRoomy } from "../roomy/to.js";
+import { fingerprint } from "../utils/hash.js";
+import { getRoomKey } from "../utils/room.js";
 
 /**
  * Compute a SHA-256 hash of normalized Discord message content.
